+++ src/main/java/kr/hhplus/be/ecommerce/interfaces/order/api/OrderController.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.ecommerce.interfaces.order;
+package kr.hhplus.be.ecommerce.interfaces.order.api;
 
 import jakarta.validation.Valid;
 import kr.hhplus.be.ecommerce.application.order.OrderFacade;
